requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
simple-salesforce==1.12.6
python-dotenv==1.0.1
black==24.2.0
isort==5.13.2
pre-commit==3.6.0
pyyaml==6.0.1
